+++ bin/nanocld/src/tasks/vm.rs
@@ -64,7 +64,9 @@ impl ObjTaskDelete for VmDb {
       .await?;
       utils::vm_image::delete_by_pk(&vm.spec.disk.image, &state).await?;
       VmDb::clear_by_pk(&vm.spec.vm_key, &state.inner.pool).await?;
-      state.emit_normal_native_action(&vm, NativeEventAction::Destroy);
+      state
+        .emit_normal_native_action_sync(&vm, NativeEventAction::Destroy)
+        .await;
       Ok::<_, IoError>(())
     })
   }